+++ modules/actor/actor-sheet.js
@@ -237,7 +237,7 @@ export class HMActorSheet extends foundry.appv1.sheets.ActorSheet {
             const handler = ev => {
                 try {
                     this._onDragStart(ev);
-                } catch (error) {
+                } catch(error) {
                     if (!(error instanceof TypeError)) throw error;
                     HMContainer.dragStartHandler(ev, this.actor);
                 }
@@ -388,7 +388,7 @@ export class HMActorSheet extends foundry.appv1.sheets.ActorSheet {
         let cardType = false;
 
         if (dialog === "save") return actor.rollSave(dataset);
-        if (dialog === "ability") cardType = CHAT_TYPE.ABILITY_CHECK;
+        if (dialog === "ability") return actor.rollAbility(dataset);
 
         if (dialog === "atk") {
             return game[SYSTEM_ID].HMWeaponItem.rollAttack({ weapon: dataset.itemId, caller: actor });