+++ modules/dialog/dialog-factory.js
@@ -3,7 +3,6 @@ import { HMDialog } from "./dialog.js";
 
 export const HMDialogFactory = (dataset, caller = null, opt = {}) => {
     const { dialog } = dataset;
-    if (dialog === "ability") return HMDialog.getAbilityDialog(dataset, caller);
     if (dialog === "atk") return HMDialog.getAttackDialog(dataset, caller, opt);
     if (dialog === "cast") return HMDialog.getCastDialog(dataset, caller, opt);
     if (dialog === "def") return HMDialog.getDefendDialog(dataset, caller);