+++ modules/sys/templates.js
@@ -70,7 +70,7 @@ export const preloadHandlebarsTemplates = () => {
         systemPath("templates/item/parts/header-logistics.hbs"),
         systemPath("templates/item/parts/header.hbs"),
 
-        systemPath("templates/dialog/getAbility.hbs"),
+        systemPath("templates/dialog/ability-content-app.hbs"),
         systemPath("templates/dialog/getAttack.hbs"),
         systemPath("templates/dialog/cast.hbs"),
         systemPath("templates/dialog/critical-content-app.hbs"),